pyproject.toml
[project]
name = "autofxxk"
version = "0.1.0"
description = "金融分析多 Agent 系统，基于 Agno 构建"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "yfinance>=0.2.0",
    "duckduckgo-search>=8.0.0",
    "pyyaml>=6.0.0",
    "ddgs>=9.10.0",
]
